Various improvements - post workshop (#201)
Complete refactor of all models to support SQLAlchemy 2
Switch from gunicorn to granian


Fixes #135
Fixes #136 


Breaking change:

Story & NewsItem IDs are now str
NewsItemData removed, all newsitem data now lives directly inside NewsItem

+++ src/core/core/model/asset.py
@@ -1,42 +1,36 @@
 import uuid
 from sqlalchemy import or_
+from sqlalchemy.sql import Select
+from sqlalchemy.orm import Mapped, relationship
 from typing import Any
 
 from core.managers.db_manager import db
 from core.model.base_model import BaseModel
 from core.model.report_item import ReportItem
-from core.model.user import User
 from core.model.organization import Organization
 
 
-class AssetCpe(BaseModel):
-    id = db.Column(db.Integer, primary_key=True)
-    value: Any = db.Column(db.String())
-
-    asset_id = db.Column(db.Integer, db.ForeignKey("asset.id"))
-    asset = db.relationship("Asset")
-
-    def __init__(self, value):
-        self.id = None
-        self.value = value
-
-
 class Asset(BaseModel):
-    id = db.Column(db.Integer, primary_key=True)
-    name: Any = db.Column(db.String(), nullable=False)
-    serial: Any = db.Column(db.String())
-    description: Any = db.Column(db.String())
+    __tablename__ = "asset"
 
-    asset_group_id: Any = db.Column(db.String, db.ForeignKey("asset_group.id"))
-    asset_group: Any = db.relationship("AssetGroup")
+    id: Mapped[int] = db.Column(db.Integer, primary_key=True)
+    name: Mapped[str] = db.Column(db.String(), nullable=False)
+    serial: Mapped[str] = db.Column(db.String())
+    description: Mapped[str] = db.Column(db.String())
 
-    asset_cpes: Any = db.relationship("AssetCpe", cascade="all, delete-orphan", back_populates="asset")
+    asset_group_id: Mapped[str] = db.Column(db.String, db.ForeignKey("asset_group.id"))
+    asset_group: Mapped["AssetGroup"] = relationship("AssetGroup")
 
-    vulnerabilities: Any = db.relationship("AssetVulnerability", cascade="all, delete-orphan", back_populates="asset")
-    vulnerabilities_count = db.Column(db.Integer, default=0)
+    asset_cpes: Mapped[list["AssetCpe"]] = relationship("AssetCpe", cascade="all, delete-orphan", back_populates="asset")
+
+    vulnerabilities: Mapped[list["AssetVulnerability"]] = relationship(
+        "AssetVulnerability", cascade="all, delete-orphan", back_populates="asset"
+    )
+    vulnerabilities_count: Mapped[int] = db.Column(db.Integer, default=0)
 
     def __init__(self, name, serial, description, group, asset_cpes=None, id=None):
-        self.id = id
+        if id:
+            self.id = id
         self.name = name
         self.serial = serial
         self.description = description
@@ -48,19 +42,17 @@ class Asset(BaseModel):
         if len(cpes) <= 0:
             return []
 
-        return (
-            db.session.query(cls)
-            .join(AssetCpe, cls.id == AssetCpe.asset_id)
-            .filter(or_(*[AssetCpe.value.like(cpe) for cpe in cpes]))
-            .distinct()
-            .all()
+        return cls.get_filtered(
+            db.select(cls).join(AssetCpe, cls.id == AssetCpe.asset_id).filter(db.or_(*[AssetCpe.value.like(cpe) for cpe in cpes])).distinct()
         )
 
     @classmethod
     def remove_vulnerability(cls, report_item_id):
         vulnerabilities = AssetVulnerability.get_by_report(report_item_id)
+        if not vulnerabilities:
+            return
         for vulnerability in vulnerabilities:
-            vulnerability.asset.vulnerabilities_count -= 1
+            vulnerability.asset.vulnerabilities_count -= 1  # type: ignore
             db.session.delete(vulnerability)
 
     def add_vulnerability(self, report_item):
@@ -75,6 +67,8 @@ class Asset(BaseModel):
     def update_vulnerabilities(self):
         cpes = [cpe.value for cpe in self.asset_cpes if cpe]
         report_item_ids = ReportItem.get_by_cpe(cpes)
+        if not report_item_ids:
+            return
 
         solved = [vulnerability.report_item_id for vulnerability in self.vulnerabilities if vulnerability.solved is True]
         self.vulnerabilities = []
@@ -88,31 +82,29 @@ class Asset(BaseModel):
             self.vulnerabilities.append(vulnerability)
 
     @classmethod
-    def solve_vulnerability(cls, user, asset_id, report_item_id, solved):
-        asset = cls.query.get(asset_id)
-        if AssetGroup.access_allowed(user, asset.asset_group_id):
-            for vulnerability in asset.vulnerabilities:
-                if vulnerability.report_item_id == report_item_id:
-                    if solved is not vulnerability.solved:
-                        if solved:
-                            asset.vulnerabilities_count -= 1
-                        else:
-                            asset.vulnerabilities_count += 1
-                    vulnerability.solved = solved
-                    db.session.commit()
-                    return
+    def solve_vulnerability(cls, organization: Organization, asset_id, report_item_id, solved):
+        asset = cls.get(asset_id)
+        if not asset:
+            return {"error": "Asset Not Found"}, 404
+        if not AssetGroup.access_allowed(organization, asset.asset_group_id):
+            return {"error": "Access Denied"}, 403
+
+        for vulnerability in asset.vulnerabilities:
+            if vulnerability.report_item_id == report_item_id:
+                if solved is not vulnerability.solved:
+                    if solved:
+                        asset.vulnerabilities_count -= 1
+                    else:
+                        asset.vulnerabilities_count += 1
+                vulnerability.solved = solved
+                db.session.commit()
+                return
 
     @classmethod
-    def get_by_filter(cls, group_id, search, sort, vulnerable, organization):
-        query = cls.query
+    def get_filter_query(cls, filter_args: dict) -> Select:
+        query = db.select(cls)
 
-        if group_id:
-            query = query.filter(Asset.asset_group_id == group_id)
-
-        if vulnerable:
-            query = query.filter(Asset.vulnerabilities_count > 0)
-
-        if search:
+        if search := filter_args.get("search"):
             query = query.join(AssetCpe, Asset.id == AssetCpe.asset_id).filter(
                 or_(
                     Asset.name.ilike(f"%{search}%"),
@@ -122,59 +114,43 @@ class Asset(BaseModel):
                 )
             )
 
-        if organization:
+        if group_id := filter_args.get("group"):
+            query = query.filter(Asset.asset_group_id == group_id)
+
+        if vulnerable := filter_args.get("vulnerable"):
+            if vulnerable == "true":
+                query = query.filter(Asset.vulnerabilities_count > 0)
+            else:
+                query = query.filter(Asset.vulnerabilities_count == 0)
+
+        if organization := filter_args.get("organization"):
             query = query.join(AssetGroup, Asset.asset_group_id == AssetGroup.id).filter(AssetGroup.organization == organization)
 
-        if sort:
+        if sort := filter_args.get("sort"):
             if sort == "ALPHABETICAL":
                 query = query.order_by(db.asc(Asset.name))
             else:
                 query = query.order_by(db.desc(Asset.vulnerabilities_count))
 
-        return query.all()
-
-    @classmethod
-    def get_by_id(cls, asset_id, organization: Organization | None = None):
-        query = cls.query.filter(Asset.id == asset_id)
-
-        if organization:
-            query = query.join(AssetGroup, Asset.asset_group_id == AssetGroup.id).filter(AssetGroup.organization == organization)
-
-        return query.first()
-
-    @classmethod
-    def get_all_json(cls, user, filter):
-        group_id = filter.get("group")
-        search = filter.get("search")
-        sort = filter.get("sort")
-        vulnerable = filter.get("vulnerable")
-        organization = user.organization
-        assets = cls.get_by_filter(group_id, search, sort, vulnerable, organization)
-        return [asset.to_dict() for asset in assets], 200
-
-    @classmethod
-    def load_multiple(cls, json_data: list[dict[str, Any]]) -> list["Asset"]:
-        return [cls.from_dict(data) for data in json_data]
-
-    @classmethod
-    def from_dict(cls, data: dict[str, Any]) -> "Asset":
-        return cls(**data)
+        return query
 
     def to_dict(self):
-        data = {c.name: getattr(self, c.name) for c in self.__table__.columns}
+        data = super().to_dict()
         data["asset_cpes"] = [asset_cpe.id for asset_cpe in self.asset_cpes if asset_cpe]
         data["vulnerabilities"] = [vulnerability.id for vulnerability in self.vulnerabilities]
         return data
 
     @classmethod
-    def get_json(cls, organization, asset_id):
-        asset = cls.get_by_id(asset_id, organization)
-        return (asset.to_dict(), 200) if asset else ("Asset Not Found", 404)
+    def get_for_api(cls, item_id, organization: Organization) -> tuple[dict[str, Any], int]:
+        if item := cls.get(item_id):
+            if AssetGroup.access_allowed(organization, item.asset_group_id):
+                return item.to_dict(), 200
+        return {"error": f"{cls.__name__} {item_id} not found"}, 404
 
     @classmethod
-    def add(cls, user: User, data) -> tuple[dict, int]:
+    def add(cls, organization: Organization, data) -> tuple[dict, int]:
         asset = cls.from_dict(data)
-        if not AssetGroup.access_allowed(user, asset.asset_group_id):
+        if not AssetGroup.access_allowed(organization, asset.asset_group_id):
             return {"error": "Access Denied"}, 403
 
         db.session.add(asset)
@@ -183,12 +159,12 @@ class Asset(BaseModel):
         return {"message": "Asset added", "id": asset.id}, 201
 
     @classmethod
-    def update(cls, user, asset_id, data) -> tuple[dict, int]:
-        asset = cls.query.get(asset_id)
+    def update(cls, organization: Organization, asset_id, data) -> tuple[dict, int]:
+        asset = cls.get(asset_id)
         if not asset:
             return {"error": "Asset Not Found"}, 404
 
-        if not AssetGroup.access_allowed(user, asset.asset_group_id):
+        if not AssetGroup.access_allowed(organization, asset.asset_group_id):
             return {"error": "Access Denied"}, 403
         for key, value in data.items():
             if hasattr(asset, key) and key != "id":
@@ -198,12 +174,12 @@ class Asset(BaseModel):
         return {"message": "Asset updated", "id": asset.id}, 201
 
     @classmethod
-    def delete(cls, user, id) -> tuple[dict, int]:
-        asset = cls.query.get(id)
+    def delete(cls, organization: Organization, id) -> tuple[dict, int]:
+        asset = cls.get(id)
         if not asset:
             return {"error": "Asset Not Found"}, 404
 
-        if not AssetGroup.access_allowed(user, asset.asset_group_id):
+        if not AssetGroup.access_allowed(organization, asset.asset_group_id):
             return {"error": "Access Denied"}, 403
 
         db.session.delete(asset)
@@ -212,32 +188,35 @@ class Asset(BaseModel):
 
 
 class AssetVulnerability(BaseModel):
-    id = db.Column(db.Integer, primary_key=True)
-    solved = db.Column(db.Boolean, default=False)
+    __tablename__ = "asset_vulnerability"
 
-    asset_id = db.Column(db.Integer, db.ForeignKey("asset.id"))
-    asset = db.relationship("Asset", back_populates="vulnerabilities")
+    id: Mapped[int] = db.Column(db.Integer, primary_key=True)
+    solved: Mapped[bool] = db.Column(db.Boolean, default=False)
 
-    report_item_id = db.Column(db.String(64), db.ForeignKey("report_item.id", ondelete="CASCADE"))
-    report_item = db.relationship("ReportItem")
+    asset_id: Mapped[int] = db.Column(db.Integer, db.ForeignKey("asset.id"))
+    asset: Mapped["Asset"] = relationship("Asset", back_populates="vulnerabilities")
+
+    report_item_id: Mapped[str] = db.Column(db.String(64), db.ForeignKey("report_item.id", ondelete="CASCADE"))
+    report_item: Mapped["ReportItem"] = relationship("ReportItem")
 
     def __init__(self, asset_id, report_item_id):
-        self.id = None
         self.asset_id = asset_id
         self.report_item_id = report_item_id
 
     @classmethod
     def get_by_report(cls, report_id):
-        return cls.query.filter_by(report_item_id=report_id).all()
+        return cls.get_filtered(db.select(cls).filter_by(report_item_id=report_id))
 
 
 class AssetGroup(BaseModel):
-    id = db.Column(db.String(64), primary_key=True)
-    name: Any = db.Column(db.String(), nullable=False)
-    description: Any = db.Column(db.String())
+    __tablename__ = "asset_group"
 
-    organization_id = db.Column(db.Integer, db.ForeignKey("organization.id"))
-    organization: Any = db.relationship("Organization")
+    id: Mapped[str] = db.Column(db.String(64), primary_key=True)
+    name: Mapped[str] = db.Column(db.String(), nullable=False)
+    description: Mapped[str] = db.Column(db.String())
+
+    organization_id: Mapped[int] = db.Column(db.Integer, db.ForeignKey("organization.id"))
+    organization: Mapped["Organization"] = relationship("Organization")
 
     def __init__(self, name, description, organization, id=None):
         self.id = id or str(uuid.uuid4())
@@ -246,21 +225,29 @@ class AssetGroup(BaseModel):
         self.organization = Organization.get(organization) if isinstance(organization, int) else organization
 
     @classmethod
-    def access_allowed(cls, user: User, group_id: str):
-        return cls.query.get(group_id).organization == user.organization
+    def access_allowed(cls, organization: Organization, group_id: str) -> bool:
+        if group_id == "default":
+            return True
+        if group := cls.get(group_id):
+            return group.organization == organization
+        return False
 
     @classmethod
     def get_default_group(cls):
-        return cls.query.get("default")
+        return cls.get("default")
 
     @classmethod
-    def get_by_filter(cls, search: str | None, organization: Organization | None = None):
-        query = cls.query
+    def get_for_api(cls, item_id, organization: Organization) -> tuple[dict[str, Any], int]:
+        if not cls.access_allowed(organization, item_id):
+            return {"error": "Access denied"}, 403
 
-        if organization:
-            query = query.filter_by(organization_id=organization.id)
+        return super().get_for_api(item_id)
 
-        if search:
+    @classmethod
+    def get_filter_query(cls, filter_args: dict) -> Select:
+        query = db.select(cls)
+
+        if search := filter_args.get("search"):
             query = query.filter(
                 or_(
                     AssetGroup.name.ilike(f"%{search}%"),
@@ -268,36 +255,49 @@ class AssetGroup(BaseModel):
                 )
             )
 
-        return query.order_by(db.asc(AssetGroup.name)).all(), query.count()
+        if organization := filter_args.get("organization"):
+            query = query.filter_by(organization_id=organization.id)
+
+        return query.order_by(db.asc(AssetGroup.name))
 
     @classmethod
-    def get_all_json(cls, user, search):
-        groups, count = cls.get_by_filter(search, user.organization)
-        return [group.to_dict() for group in groups]
-
-    @classmethod
-    def delete(cls, user, group_id):
+    def delete(cls, organization: Organization, group_id):
         if group_id == "default":
-            return "Cannot delete default group", 400
-        if not cls.access_allowed(user, group_id):
-            return "Access denied", 403
+            return {"error": "Cannot delete default group"}, 400
+        if not cls.access_allowed(organization, group_id):
+            return {"error": "Access denied"}, 403
 
-        group = cls.query.get(group_id)
+        group = cls.get(group_id)
         db.session.delete(group)
         db.session.commit()
 
-        return "Group deleted", 200
+        return {"message": "Group deleted"}, 200
 
     @classmethod
-    def update(cls, user, group_id, data):
-        if not cls.access_allowed(user, group_id):
-            return "Access denied", 403
+    def update(cls, organization: Organization, group_id, data):
+        if not cls.access_allowed(organization, group_id):
+            return {"error": "Access denied"}, 403
 
-        updated_group = cls.from_dict(data)
-        if not updated_group:
-            return "Invalid group data", 400
-        group = cls.query.get(group_id)
-        group.name = updated_group.name
-        group.description = updated_group.description
+        group = cls.get(group_id)
+        if not group:
+            return {"error": "Group not found"}, 404
+
+        if name := data.get("name"):
+            group.name = name
+        if description := data.get("description"):
+            group.description = description
         db.session.commit()
-        return "Group updated", 200
+        return {"message": "Group updated"}, 200
+
+
+class AssetCpe(BaseModel):
+    __tablename__ = "asset_cpe"
+
+    id: Mapped[int] = db.Column(db.Integer, primary_key=True)
+    value: Mapped[str] = db.Column(db.String())
+
+    asset_id = db.Column(db.Integer, db.ForeignKey("asset.id"))
+    asset = db.relationship("Asset")
+
+    def __init__(self, value):
+        self.value = value
